requirements.txt
moviepy>=1.0.3
python-ffmpeg>=2.0.0
pillow==9.0.1
openai
pydantic
python-dotenv
markdown2
streamlit
elevenlabs
opencv-python